Added nginx configuration

+++ docker/vm_startup.sh
@@ -1,5 +1,14 @@
 #!/bin/bash
 # Setup VM
+echo "=============="
+echo "SCRIPT STARTING"
+echo "=============="
+
+if [[ ! -e backend.env || ! -e frontend.env]]
+then
+    echo "Necessary env files are missing. Please create it first before running this script."
+    exit 1
+fi
 
 echo "Downloading Docker"
 curl -fsSL https://get.docker.com -o install-docker.sh
@@ -9,3 +18,13 @@ sudo sh install-docker.sh
 
 echo "Adding user into docker group"
 sudo usermod -aG docker $USER
+
+echo "Creating the containers"
+sudo docker compose -f docker-compose.yaml up -d
+
+echo "Seeding the database"
+sudo docker exec -it synapse_backend bash
+
+echo "=============="
+echo "SCRIPT FINISHED"
+echo "=============="
